devops(consents): refactor to model (#7998)
* Model consents base

* '@blockframes/consents/+state/consents.firestore' to @blockframes/model

* './consents.firestore' to @blockframes/model
