+++ libs/consents/src/lib/+state/consents.service.ts
@@ -1,8 +1,7 @@
 import { Injectable } from '@angular/core';
 import { AngularFireFunctions } from '@angular/fire/functions';
 import { IpService } from '@blockframes/utils/ip';
-import { ConsentType } from './consents.firestore'
-
+import { ConsentType } from '@blockframes/model';
 
 @Injectable({ providedIn: 'root' })
 export class ConsentsService {
